requirements.txt
streamlit
transformers
beautifulsoup4
feedparser
langchain
faiss-cpu
sentence-transformers
unstructured
python-dotenv
